RaspbotController/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
smbus2==0.5.0
python-multipart==0.0.12
pydantic-settings==2.5.2
